+++ CHANGELOG.md
@@ -7,7 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 n/a
 
+## [1.0.2] - 2018-01-19
+### Fixed
+- Fixed static instance warning
+- Fixed failure of ACF function by including admin plugin.php 
+
+### Updates
+- Updates all functions named bwe_ to bre_
+
 ## [1.0.1] - 2018-01-19
 ### Added
 - Changelog
-- readme.txt for WP repo 
+- readme.txt for WP repo